requirements.txt
ossapi == 3.3.2
python-dotenv == 1.0.0
requests == 2.31.0